refactor: 메인화면 컴포넌트 파일구조 변경
