+++ today's horoscope/apps/react-web/src/pages/home/Home.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import './Home.scss';
-import Carousel from './components/Carousel';
+import Carousel from './components/Carousel/Carousel';
 import { IoMenuOutline } from 'react-icons/io5';
-import MenuModal from './components/MenuModal';
+import MenuModal from './components/MenuModal/MenuModal';
 
 function Home() {
   const [menuModal, setMenuModal] = useState<boolean>(false);
